Implement Phase 4: MVC Architecture Refactoring
Replace CircuitDesignGUI with MainWindow using proper MVC pattern.

Changes:
- Create MainWindow class that instantiates model and controllers
- Delegate file operations to FileController
- Delegate simulation to SimulationController
- Delegate circuit operations to CircuitController
- Add canvas sync methods (sync_to_model, sync_from_model)
- Remove old CircuitDesignGUI class
- Update all imports to use MainWindow

Architecture:
- Model (CircuitModel) is framework-agnostic (no Qt)
- Controllers share single model reference
- Controllers don't import Qt (MVC separation maintained)
- MainWindow handles only UI construction and view coordination
