+++ app/__init__.py
@@ -1,4 +1,4 @@
-from .GUI import (AnalysisDialog, CircuitCanvasView, CircuitCanvas, CircuitDesignGUI, Node,
+from .GUI import (AnalysisDialog, CircuitCanvasView, CircuitCanvas, MainWindow, Node,
                  ComponentGraphicsItem, ComponentPalette, #GridPathfinder,
                  get_component_obstacles, WireGraphicsItem, WireItem)
 from .simulation import NetlistGenerator, NgspiceRunner, ResultParser
@@ -7,7 +7,7 @@ __all__ = ['NetlistGenerator', 'NgspiceRunner', 'ResultParser',
     'AnalysisDialog',
     'CircuitCanvasView',
     'CircuitCanvas',  # Backward compatibility
-    'CircuitDesignGUI',
+    'MainWindow',
     'Node',
     'ComponentGraphicsItem',
     'ComponentPalette',
